sem_patch_extractor/pyproject.toml
[project]
name = "sem_patch_extractor"
version = "0.1.0"
description = "Extract N µm × N µm patches from SEM .tif images using scalebar or metadata"
authors = [
  { name = "Andrea Cicconardi", email = "andrea.cicconardi@iit.it" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "numpy",
  "Pillow",
  "scipy",
  "matplotlib"
]

[project.urls]
Repository = "https://github.com/andreacicconardi94/Research-Work"
